+++ src/cljs/bundled/lumo/compiler.cljs
@@ -180,50 +180,8 @@
         source
         (:ns ns-info)
         (merge opts {:verbose false
-                     :load (fn load
-                             [{:keys [macros path]
-                               dep :name :as m} cb]
-                             (let [cenv @env/*compiler*
-                                   dep (if macros
-                                         (symbol (str dep "$macros"))
-                                         dep)
-                                   cache (get-in cenv [::ana/namespaces dep :defs])]
-                               (if (or (not-empty cache)
-                                       (contains? (:js-dependency-index cenv) (name dep))
-                                       (ana/node-module-dep? dep)
-                                       (ana/js-module-exists? (name dep))
-                                       (deps/find-classpath-lib dep))
-                                 (cb {:lang :js})
-                                 (if-let [{:keys [source-file]} (first
-                                                                  (filter
-                                                                    #(symbol-identical? dep (:ns %))
-                                                                    (:sources cenv)))]
-                                   (cb {:lang :clj
-                                        :file (util/path source-file)
-                                        :source (slurp source-file)})
-                                   (if (and (not macros) (empty? cache))
-                                     (let [path (path/join (util/output-directory opts) path)
-                                           f (or (io/resource (str path ".cljs"))
-                                                 (io/resource (str path ".cljc")))]
-                                       (try
-                                         (lana/read-analysis-cache
-                                           (when (:cache-analysis opts)
-                                             (lana/cache-file f (lana/parse-ns dep) (util/output-directory opts)))
-                                           f
-                                           opts)
-                                         (let [cache (get-in @env/*compiler* [::ana/namespaces dep])]
-                                           (cljs/analyze-deps {:*compiler* env/*compiler*
-                                                               :*cljs-dep-set* ana/*cljs-dep-set*
-                                                               :*load-fn* load}
-                                             nil
-                                             dep
-                                             (distinct (concat (vals (:requires cache)) (vals (:imports cache))))
-                                             opts
-                                             cb)
-                                           (cb {:lang :js}))
-                                         (catch :default e
-                                           (original-load m cb))))
-                                     (original-load m cb))))))})
+                     :load (fn [m cb]
+                             (lana/compiler-load m opts original-load cb))})
         #(vreset! cb-val %)))
     (let [{:keys [value error]} @cb-val]
       (if error